+++ adminforth/spa/src/renderers/CompactUUID.vue
@@ -1,11 +1,11 @@
 <template>
   <span class="flex items-center"  
-    :data-tooltip-target="`tooltip-${id}`"
+    :data-tooltip-target="val && `tooltip-${id}`"
     data-tooltip-placement="top"
   >
-    {{ visualValue }} <IconFileCopyAltSolid @click.stop="copyToCB" class="w-5 h-5 text-lightPrimary dark:text-darkPrimary"/>
+    {{ visualValue }} <IconFileCopyAltSolid @click.stop="copyToCB" class="w-5 h-5 text-lightPrimary dark:text-darkPrimary" v-if="val"/>
 
-    <div :id="`tooltip-${id}`" role="tooltip"
+    <div :id="`tooltip-${id}`" role="tooltip" v-if="val"
           class="absolute z-10 invisible inline-block px-3 py-2 text-sm font-medium text-white transition-opacity duration-300 bg-gray-900 rounded-lg shadow-sm opacity-0 tooltip dark:bg-gray-700">
       {{ props.record[props.column.name] }}
       <div class="tooltip-arrow" data-popper-arrow></div>